Add files via upload
Welcome to the Calculator Web App project! This is my first internship project, developed under Techno Hacks, where I had the opportunity to dive into the world of web development. In this project, I created a simple yet functional calculator using HTML, CSS, and JavaScript.

Technologies Used:
HTML: The foundation of the project, providing the structure and layout of the calculator interface.

CSS: To style and make the calculator visually appealing, ensuring a user-friendly experience.

JavaScript: The core of the calculator's functionality, enabling arithmetic operations and user interactions.