Add files via upload
Welcome to the Temperature Converter Web App project! This is my second internship project, developed under Techno Hacks, where I continued to explore the world of web development. In this project, I created a versatile temperature converter using HTML, CSS, and JavaScript.

Technologies Used:
HTML: The backbone of the project, providing the structure and layout for the temperature converter interface.

CSS: To style and enhance the user interface, ensuring an intuitive and visually appealing experience.

JavaScript: The heart of the converter's functionality, allowing users to convert between different temperature units.