Add files via upload
Welcome to the Registration Page Web Design project! This is my third internship project, developed under Techno Hacks, where I've had the opportunity to delve deeper into web development. In this project, I've created an elegant and user-friendly registration page using HTML and CSS.

Technologies Used:
HTML: The foundation of the project, providing the structure and layout for the registration page.

CSS: To style and beautify the user interface, ensuring a polished and intuitive registration experience.